Done file partitioning of gql models
