+++ server/gql/models/studentSubjects/customUpdateResolver.js
@@ -1,9 +1,11 @@
 import { updateStudent } from '@server/daos/students';
 import { updateStudentSubject } from '@server/daos/studentSubjects';
 import { updateSubject } from '@server/daos/subjects';
+import db from '@server/database/models';
 import { transformSQLError } from '@server/utils';
 
 export const customUpdateResolver = async (model, args, context) => {
+  const t = await db.sequelize.transaction();
   try {
     if (args?.studentName) {
       const studentArgs = {
@@ -11,7 +13,7 @@ export const customUpdateResolver = async (model, args, context) => {
         name: args.studentName
       };
       delete args.studentName;
-      await updateStudent(studentArgs);
+      await updateStudent(studentArgs, { transaction: t });
     }
     if (args?.subjectName) {
       const subjectArgs = {
@@ -19,10 +21,13 @@ export const customUpdateResolver = async (model, args, context) => {
         name: args.subjectName
       };
       delete args.subjectName;
-      await updateSubject(subjectArgs);
+      await updateSubject(subjectArgs, { transaction: t });
     }
-    return updateStudentSubject(args);
+    const res = updateStudentSubject(args, { transaction: t });
+    await t.commit();
+    return res;
   } catch (err) {
+    await t.rollback();
     return transformSQLError(err);
   }
 };
