+++ server/gql/models/students/index.js
@@ -1,39 +1,35 @@
+import { student } from './query';
+import { studentConnection } from './list';
+import { GraphQLInt, GraphQLNonNull } from 'graphql';
+import { limitAndOffset } from '@server/gql/fields/limitAndOffset';
+import db from '@server/database/models';
 import { studentMutation } from './mutation';
-import { GraphQLID, GraphQLNonNull, GraphQLObjectType, GraphQLString } from 'graphql';
-import { getNode } from '@server/gql/node';
-import { getQueryFields, TYPE_ATTRIBUTES } from '@server/utils/gqlFieldUtils';
-import { timestamps } from '@server/gql/fields/timestamps';
-import { subjectQueries } from '../subjects';
-import { StudentConnection, studentQuery } from './query';
 
-// imported GraphQLObjectType of student.
-// const Student = student;
+export const Student = student;
 
-const { nodeInterface } = getNode();
+// exporting student connection.
+export const StudentConnection = studentConnection;
 
-export const studentsFields = {
-  id: { type: new GraphQLNonNull(GraphQLID) },
-  name: { type: GraphQLString }
-};
-
-const Student = new GraphQLObjectType({
-  name: 'Student',
-  interfaces: [nodeInterface],
-  fields: () => ({
-    ...getQueryFields(studentsFields, TYPE_ATTRIBUTES.isNonNull),
-    ...timestamps,
-    subjects: {
-      ...subjectQueries.list,
-      resolve: (source, args, context, info) =>
-        subjectQueries.list.resolve(source, args, { ...context, student: source.dataValues }, info)
+export const studentQueries = {
+  args: {
+    id: {
+      type: new GraphQLNonNull(GraphQLInt)
     }
-  })
-});
-
-export { StudentConnection, Student };
-
-// exporting queries on the students table.
-export const studentQueries = studentQuery;
+  },
+  query: {
+    type: student
+  },
+  list: {
+    ...StudentConnection,
+    resolve: StudentConnection.resolve,
+    type: StudentConnection.connectionType,
+    args: {
+      ...StudentConnection.connectionArgs,
+      ...limitAndOffset
+    }
+  },
+  model: db.students
+};
 
 // exporting mutations on the students table.
 export const studentMutations = studentMutation;
