+++ server/gql/models/students/list.js
@@ -1,25 +1,26 @@
 import db from '@database/models';
-import { limitAndOffset } from '@server/gql/fields/limitAndOffset';
-import { GraphQLInt, GraphQLNonNull } from 'graphql';
-import { Student, StudentConnection } from '.';
+import { student } from './query';
+import { sequelizedWhere } from '@server/database/dbUtils';
+import { totalConnectionFields } from '@server/utils';
+import { createConnection } from 'graphql-sequelize';
 
-export const studentQuery = {
-  args: {
-    id: {
-      type: new GraphQLNonNull(GraphQLInt)
+export const studentConnection = createConnection({
+  name: 'students',
+  target: db.students,
+  nodeType: student,
+  before: (findOptions, args, context) => {
+    findOptions.include = findOptions.include || [];
+
+    if (context?.subject?.id) {
+      findOptions.include.push({
+        model: db.studentSubjects,
+        where: {
+          subjectId: context.subject.id
+        }
+      });
     }
+    findOptions.where = sequelizedWhere(findOptions.where, args.where);
+    return findOptions;
   },
-  query: {
-    type: Student
-  },
-  list: {
-    ...StudentConnection,
-    resolve: StudentConnection.resolve,
-    type: StudentConnection.connectionType,
-    args: {
-      ...StudentConnection.connectionArgs,
-      ...limitAndOffset
-    }
-  },
-  model: db.students
-};
+  ...totalConnectionFields
+});
